+++ src/App.js
@@ -1,34 +1,15 @@
-
+import React, { Component } from 'react'
 import './App.css';
-import IngredientList from './IngredientList';
-import BurgerPane from './BurgerPane';
+
+
 
 function App() {
-  const ingredients = [
-    {name: 'Kaiser Bun', color: 'saddlebrown'},
-    {name: 'Sesame Bun', color: 'sandybrown'},
-    {name: 'Gluten Free Bun', color: 'peru'},
-    {name: 'Lettuce Wrap', color: 'olivedrab'},
-    {name: 'Beef Patty', color: '#3F250B'},
-    {name: 'Soy Patty', color: '#3F250B'},
-    {name: 'Black Bean Patty', color: '#3F250B'},
-    {name: 'Chicken Patty', color: 'burlywood'},
-    {name: 'Lettuce', color: 'lawngreen'},
-    {name: 'Tomato', color: 'tomato'},
-    {name: 'Bacon', color: 'maroon'},
-    {name: 'Onion', color: 'lightyellow'}
-]
+  
   return (
     <div className="App">
       <header className="App-header">
-        <div className="flex-container">
-          
-              <IngredientList />
-            
-            
-              <BurgerPane />
-            
-        </div>
+        
+
       </header>
     </div>
   );
