+++ src/BurgerPane.js
@@ -4,7 +4,7 @@ export default class BurgerPane extends Component {
     render() {
         return(
             <div className="container">
-                Container 2
+                Burger Panels
             </div>
         )
     }
