+++ src/IngredientList.js
@@ -1,13 +1,16 @@
 import React, { Component } from 'react'
+import Ingredients from './Ingredients'
 
-export default class Ingredients extends Component {
-
-    
-
+export default class IngredientList extends Component {
+    state = {
+        ingredientStuff: this.props.ingredientsBook
+    }
     render() {
+        
         return(
             <div className="container">
-                Container 1
+                Ingredients List:
+                <Ingredients ingredientsToShow = {this.state.ingredientStuff}/>
             </div>
         )
