+++ src/index.js
@@ -1,12 +1,32 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
 import reportWebVitals from './reportWebVitals';
+import IngredientList from './IngredientList';
+import BurgerPane from './BurgerPane';
+
+const ingredientsOfThings = [
+  {name: 'Kaiser Bun', color: 'saddlebrown'},
+  {name: 'Sesame Bun', color: 'sandybrown'},
+  {name: 'Gluten Free Bun', color: 'peru'},
+  {name: 'Lettuce Wrap', color: 'olivedrab'},
+  {name: 'Beef Patty', color: '#3F250B'},
+  {name: 'Soy Patty', color: '#3F250B'},
+  {name: 'Black Bean Patty', color: '#3F250B'},
+  {name: 'Chicken Patty', color: 'burlywood'},
+  {name: 'Lettuce', color: 'lawngreen'},
+  {name: 'Tomato', color: 'tomato'},
+  {name: 'Bacon', color: 'maroon'},
+  {name: 'Onion', color: 'lightyellow'}
+]
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <div className="flex-container">
+      <IngredientList ingredientsBook = { ingredientsOfThings } />
+      <BurgerPane />
+    </div>
+
   </React.StrictMode>,
   document.getElementById('root')
 );
